Vagrantfile: PoC cumulus networking
So the concept is we assign a loopback ip-address to a switch, and each
port the switch gets a port on that loopback ip-address. Each vm gets
a different ip-address on the loopback interface, and a port on that address
to represent the ethernet port in the vm.

spine01.swp1 = 127.0.1.1:10001
spine01.swp2 = 127.0.1.1:10002
...
controller-01.eth0 = 127.0.1.2:10000
controller-02.eth0 = 127.0.1.3:10000

https://docs.nvidia.com/networking-ethernet-software/cumulus-vx/KVM-QEMU-and-Vagrant/
https://codingpackets.com/blog/pseudo-wires-with-vagrant-and-libvirt/