Vagrantfile: more complex topology
Adding a second leaf, and secondary adapter for controllers. Then
configuring mc-lag between leaf-01 and leaf-02. And bonding for
controller-01 and controller-02 with one adapter per leaf.

This is as far as you can/should go while manually setting the values
in vagrant. Next iteration we'll introduce functions to take care
of the connections.