Vagrantfile: move additional VMs to nodes.yaml
Create VMs based on a list in nodes.yaml, introducing even more
complex maths for calculating mac addresses based on which leaf
they're connected to, which node they are, and which interface.

This numbering-scheme is mostly to easily spot connection errors,
which obviously won't happen anymore with this rock-solid code.