Workaround: don't apply settings in foreman 3.1
IMPORTANT MANUAL STEP:
This version introduces foreman_proxy::register, so you'll need to temporarily
set `create_new_host_when_facts_are_uploaded` to true.

For some reason, integers and booleans are not handled properly.

Possibly some reason the settings_type is not in the database already and
not getting set when applying settings. So values with integer as type
are interpreted as strings and foreman just crashes.

sudo -u postgres psql --dbname=foreman \
  -c "SELECT name, LEFT(value, 15) AS value, settings_type FROM settings;"